requirements.txt
fastapi>=0.68.0
uvicorn[standard]
torch>=1.13.0
torchvision
Pillow
numpy
ultralytics>=8.0.0